+++ pkg/document/crdt/rga_tree_split.go
@@ -508,10 +508,10 @@ func (s *RGATreeSplit[V]) findBetween(from, to *RGATreeSplitNode[V]) []*RGATreeS
 func (s *RGATreeSplit[V]) deleteNodes(
 	candidates []*RGATreeSplitNode[V],
 	editedAt *time.Ticket,
-	versionVector time.VersionVector,
+	vector time.VersionVector,
 ) map[string]*RGATreeSplitNode[V] {
 	removedNodeMap := make(map[string]*RGATreeSplitNode[V])
-	isVersionVectorEmpty := len(versionVector) == 0
+	isVersionVectorEmpty := len(vector) == 0
 
 	if len(candidates) == 0 {
 		return removedNodeMap
@@ -533,7 +533,7 @@ func (s *RGATreeSplit[V]) deleteNodes(
 			clientLamportAtChange = time.MaxLamport
 		} else {
 			// Case 2: from operation with version vector(After v0.5.7)
-			lamport, ok := versionVector.Get(actorID)
+			lamport, ok := vector.Get(actorID)
 			if ok {
 				clientLamportAtChange = lamport
 			} else {
@@ -577,20 +577,6 @@ func (s *RGATreeSplit[V]) deleteIndexNodes(boundaries []*RGATreeSplitNode[V]) {
 			s.treeByIndex.DeleteRange(leftBoundary.indexNode, rightBoundary.indexNode)
 		}
 	}
-}
-
-func (s *RGATreeSplit[V]) string() string {
-	builder := strings.Builder{}
-
-	node := s.initialHead.next
-	for node != nil {
-		if node.removedAt == nil {
-			builder.WriteString(node.String())
-		}
-		node = node.next
-	}
-
-	return builder.String()
 }
 
 func (s *RGATreeSplit[V]) nodes() []*RGATreeSplitNode[V] {