+++ pkg/document/operations/move.go
@@ -79,7 +79,7 @@ func (o *Move) ExecutedAt() *time.Ticket {
 }
 
 // SetActor sets the given actor to this operation.
-func (o *Move) SetActor(actorID *time.ActorID) {
+func (o *Move) SetActor(actorID time.ActorID) {
 	o.executedAt = o.executedAt.SetActorID(actorID)
 }