+++ test/helper/helper.go
@@ -77,7 +77,6 @@ var (
 	AdminTokenDuration          = "10s"
 	ClientDeactivateThreshold   = "10s"
 	SnapshotThreshold           = int64(10)
-	SnapshotWithPurgingChanges  = false
 	AuthWebhookMaxWaitInterval  = 3 * gotime.Millisecond
 	AuthWebhookMinWaitInterval  = 3 * gotime.Millisecond
 	AuthWebhookRequestTimeout   = 100 * gotime.Millisecond
@@ -152,7 +151,7 @@ func TimeT(change *change.Context) *time.Ticket {
 }
 
 // MaxVersionVector returns the max version vector of the given actors.
-func MaxVersionVector(actors ...*time.ActorID) time.VersionVector {
+func MaxVersionVector(actors ...time.ActorID) time.VersionVector {
 	if len(actors) == 0 {
 		actors = append(actors, time.InitialActorID)
 	}
@@ -166,7 +165,7 @@ func MaxVersionVector(actors ...*time.ActorID) time.VersionVector {
 }
 
 // VersionVectorOf creates a new version vector from the given actors.
-func VersionVectorOf(actors map[*time.ActorID]int64) time.VersionVector {
+func VersionVectorOf(actors map[time.ActorID]int64) time.VersionVector {
 	vector := time.NewVersionVector()
 	for actor, lamport := range actors {
 		vector.Set(actor, lamport)
@@ -282,7 +281,6 @@ func TestConfig() *server.Config {
 			ClientDeactivateThreshold:   server.DefaultClientDeactivateThreshold,
 			SnapshotInterval:            10,
 			SnapshotThreshold:           SnapshotThreshold,
-			SnapshotWithPurgingChanges:  SnapshotWithPurgingChanges,
 			AuthWebhookMaxWaitInterval:  AuthWebhookMaxWaitInterval.String(),
 			AuthWebhookMinWaitInterval:  AuthWebhookMinWaitInterval.String(),
 			AuthWebhookRequestTimeout:   AuthWebhookRequestTimeout.String(),